requirements.txt
requests==2.31.0
Flask==3.0.0
playwright==1.40.0
networkx==3.2.1
matplotlib==3.8.0
plotly==5.17.0
pandas==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3
pyOpenSSL==23.3.0
python-dotenv==1.0.0
colorama==0.4.6
fake-useragent==1.4.0
pysocks==1.7.1
cryptography==41.0.7